Add files via upload
ObjectsFloatingOneDirection features ships of the same scale to go in one horizontal direction. If the user presses the R button, the ships will go the opposite horizontal direction.

ObjectsFloatingTwoDirections features ships of the same scale to go in two random horizontal directions. If the user presses the R button, the ships will go in the opposite directions they were going before.

DifferentScaleDifferentDirections features ships of a random scale to go in random two horizontal directions. If the user presses the R button, the ships will go in the opposite directions they were going before.

In all 3 files, when the ships exit the screen, they will reappear in the opposite side.